robot_har_mln/src/requirements.txt
Flask==2.0.2
Flask_Cors==3.0.10
numpy==1.17.4
colored==1.4.3
pracmln==1.2.4
rospkg==1.3.0
rospkg_modules==1.3.0
PyYAML==5.3.1
colorama==0.4.3